extraídas constantes e a parte separável do reconhecimento de dentro da main pros próprios arquivos
